2006-03-24  Christian Stimming <stimming@tuhh.de>
        * src/os/posix/fslock.c: Cast the result of getpid() to
        int. Necessary for sparc-solaris.


git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@1010 70169cfe-8b10-0410-8925-dcb4b91034d8

+++ src/os/posix/fslock.c
@@ -309,7 +309,7 @@ int GWEN_FSLock__UnifyLockFileName(GWEN_BUFFER *nbuf) {
   GWEN_Buffer_AppendString(nbuf, "-");
 
   buffer[0]=0;
-  snprintf(buffer, sizeof(buffer)-1, "%i", getpid());
+  snprintf(buffer, sizeof(buffer)-1, "%i", (int)getpid());
   buffer[sizeof(buffer)-1]=0;
   GWEN_Buffer_AppendString(nbuf, buffer);
